pyproject.toml
[project]
name = "gemini-enrichment"
version = "0.1.0"
description = ""
authors = [
    {name = "Ryan McGuinness",email = "rmcguinness@google.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "tomlkit (>=0.13.2,<0.14.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "prettytable (>=3.14.0,<4.0.0)",
    "google-genai (>=1.0.0,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "google-cloud-trace (>=1.15.0,<2.0.0)",
    "opentelemetry-api (>=1.30.0,<2.0.0)",
    "opentelemetry-instrumentation-logging (>=0.51b0,<0.52)",
    "python-json-logger (>=3.2.1,<4.0.0)",
    "opentelemetry-exporter-otlp-proto-grpc (>=1.30.0,<2.0.0)",
    "opentelemetry-sdk (>=1.30.0,<2.0.0)",
    "opentelemetry-exporter-otlp-proto-http (>=1.30.0,<2.0.0)",
    "grpcio (>=1.70.0,<2.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "google-cloud-pubsub (>=2.28.0,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "fastapi-throttle (>=0.1.4,<0.2.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)"
]

[tool.poetry]
package-mode = true
packages = [
    { include = "cli", from = "src" },
    { include = "commands", from = "src" },
    { include = "model", from = "src" },
    { include = "utils", from = "src" },
    { include = "api", from = "src" },
    { include = "worker", from = "src" }
]

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.49"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gce = 'cli.main:main'
worker = 'worker.main:main'